Renamed cehr_mamba to ehr_mamba.
